fmap.py: sort the flags tuple
The output of this function is arbitrary because it's based on the
ordering of the dict keys of FMAP_FLAGS.  This leads to the unittest
sometimes passing & sometimes failing.  Since this is meant for us
humans, sort the tuple so it's stable and unittests are reliable.

BUG=chromium:982465
TEST=`./fmap.py bin/example.bin` works

Change-Id: Ida9e6768aa6662a400655c52bc2ba793197ab591
Reviewed-on: https://chromium-review.googlesource.com/1693887
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@google.com>